pyproject.toml
[tool.poetry]
name = "autopark"
version = "0.1.0"
description = ""
authors = ["Bearspirit <mihailvoitovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
gunicorn = "^20.1.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
